sdm660-common: init: set default wifi country code to '00'
By default, AOSP uses the 'US' country code for wifi.

When a SIM is present, Android can check the correct country (using
information from the SIM itself) and it can set the correct country code.

However, if no SIM is present or if the phone is in airplane mode,
Android will just use 'US' as country code and it will disable access to
a lot of frequencies.

To bypass this issue, set the default country code to '00': this is the
"world regulatory domain" and it contains generic settings.When a SIM is
inserted, Android will overwrite the default setting with the correct
country code.

Copy from Google:
https://android.googlesource.com/device/google/crosshatch/+/12b500667e5440345a7bd8237c38fa2839329c11/init.hardware.rc#524

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>
Signed-off-by: karthik1896 <karthik1896@outlook.com>